+++ src/cmd/troubleshoot/namespace.go
@@ -1,8 +1,6 @@
 package troubleshoot
 
 import (
-	"context"
-
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -13,7 +11,7 @@ func checkNamespace(troubleshootCtx *troubleshootContext) error {
 	logNewTestf("checking if namespace '%s' exists ...", troubleshootCtx.namespaceName)
 
 	var namespace corev1.Namespace
-	err := troubleshootCtx.apiReader.Get(context.TODO(), client.ObjectKey{Name: troubleshootCtx.namespaceName}, &namespace)
+	err := troubleshootCtx.apiReader.Get(troubleshootCtx.context, client.ObjectKey{Name: troubleshootCtx.namespaceName}, &namespace)
 
 	if err != nil {
 		return errorWithMessagef(err, "missing namespace '%s'", troubleshootCtx.namespaceName)
